pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aip-identity"
version = "0.5.11"
description = "Cryptographic identity, trust chains, and E2E encrypted messaging for AI agents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "The_Nexus_Guard_001"}
]
dependencies = [
    "pynacl>=1.5.0",
    "requests>=2.28.0",
]
keywords = [
    "ai", "agent", "identity", "trust", "cryptography", "did",
    "decentralized", "ed25519", "verification", "mcp", "llm",
    "multi-agent", "authentication", "digital-identity", "signing",
    "encrypted-messaging", "vouch", "reputation",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Distributed Computing",
]

[project.scripts]
aip = "aip_identity.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.urls]
Homepage = "https://the-nexus-guard.github.io/aip/"
Documentation = "https://aip-service.fly.dev/docs"
Repository = "https://github.com/The-Nexus-Guard/aip"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["aip_identity*"]
